--- src/pages/expand-circle/index.js
@@ -1,14 +1,0 @@
-import Canvas from '../../components/Canvas'
-
-const ExpandCircle = () => {
-  const draw = (ctx, frameCount) => {
-    ctx.clearRect(0, 0, ctx.canvas.width, ctx.canvas.height)
-    ctx.fillStyle = '#000000'
-    ctx.beginPath()
-    ctx.arc(50, 100, 20 * Math.sin(frameCount * 0.05) ** 2, 0, 2 * Math.PI)
-    ctx.fill()
-  }
-  return <Canvas draw={draw} options="2d" />
-}
-
-export default ExpandCircle
